refactor: extract store query coverage and diagnostics
